requirements.txt
django
django-allauth
djangorestframework
django-cors-headers
psycopg2
requests
jwt
djangorestframework-simplejwt